% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlm4omics-package.R
\docType{package}
\name{mlm4omics-package}
\alias{mlm4omics-package}
\alias{mlm4omics}
\title{The 'mlm4omics' package.}
\description{
To conduct Bayesian inference regression for
responses with multilevel explanatory variables and missing values
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan.
R package version 2.17.3. http://mc-stan.org
}
